convert DDS to PNG for Image Texture Node
In order to correctly render all HZD textures, images needs to be in PNG format,
as Blender can not handle all DDS formats.
Nvidia Texture Tool is needed for the conversion.
HZDEditor property added to set the full file path to nvtt_exporter.exe,
if NVTT is not installed at default path.